--- app/uk/gov/hmrc/incometaxsubscription/HelloWorld.scala
@@ -1,16 +1,0 @@
-package uk.gov.hmrc.incometaxsubscription.controllers
-
-import uk.gov.hmrc.play.frontend.controller.FrontendController
-import play.api.mvc._
-import scala.concurrent.Future
-import play.api.Play.current
-import play.api.i18n.Messages.Implicits._
-
-
-object HelloWorld extends HelloWorld
-
-trait HelloWorld extends FrontendController {
-  val helloWorld = Action.async { implicit request =>
-		Future.successful(Ok(uk.gov.hmrc.incometaxsubscription.views.html.helloworld.hello_world()))
-  }
-}
