--- app/uk/gov/hmrc/incometaxsubscription/views/helloworld/hello_world.scala.html
@@ -1,5 +1,0 @@
-@()(implicit request: Request[_], messages: Messages))
-
-@uk.gov.hmrc.incometaxsubscription.views.html.main_template(title = "Hello from income-tax-subscription-frontend", bodyClasses = None) {
-    <h1>Hello from income-tax-subscription-frontend !</h1>
-}
