+++ app/views/error_template.scala.html
@@ -8,4 +8,4 @@
   <p>@message</p>
 }
 
-@govuk_wrapper(appConfig = uk.gov.hmrc.incometaxsubscription.FrontendAppConfig, title = pageTitle, contentHeader = Some(contentHeader), mainContent = mainContent)
+@govuk_wrapper(appConfig = config.FrontendAppConfig, title = pageTitle, contentHeader = Some(contentHeader), mainContent = mainContent)
