+++ app/views/govuk_wrapper.scala.html
@@ -1,4 +1,4 @@
-@(appConfig: uk.gov.hmrc.incometaxsubscription.AppConfig,
+@(appConfig: config.AppConfig,
   title: String,
   mainClass: Option[String] = None,
   mainDataAttributes: Option[Html] = None,
