+++ app/views/helloworld/hello_world.scala.html
@@ -1,0 +1,5 @@
+@()(implicit request: Request[_], messages: Messages))
+
+@views.html.main_template(title = "Hello from income-tax-subscription-frontend", bodyClasses = None) {
+    <h1>Hello from income-tax-subscription-frontend !</h1>
+}
