+++ app/views/main_template.scala.html
@@ -17,7 +17,7 @@
 }
 
 
-@govuk_wrapper(appConfig = uk.gov.hmrc.incometaxsubscription.FrontendAppConfig,
+@govuk_wrapper(appConfig = config.FrontendAppConfig,
                title = title,
                mainClass = mainClass,
                bodyClasses = bodyClasses,
