+++ conf/app.routes
@@ -1,3 +1,3 @@
 # microservice specific routes
 
-GET        /hello-world             uk.gov.hmrc.incometaxsubscription.controllers.HelloWorld.helloWorld
+GET        /hello-world        connectors.HelloWorld.helloWorld
