+++ conf/application.conf
@@ -1,9 +1,23 @@
+# Copyright 2016 HM Revenue & Customs
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 include "common.conf"
 
 appName="income-tax-subscription-frontend"
 application.router=prod.Routes
 
-application.global=uk.gov.hmrc.incometaxsubscription.FrontendGlobal
+application.global=config.FrontendGlobal
 
 play.filters.headers.contentSecurityPolicy= "default-src 'self' 'unsafe-inline' localhost:9000 localhost:9032 www.google-analytics.com data:"
 
@@ -67,6 +81,3 @@ assets {
 contact-frontend {
   host = "http://localhost:9250"
 }
-
-
-
